backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# LLM integration
llama-cpp-python==0.2.11

# Text processing and chunking
tiktoken==0.5.1
langchain-text-splitters==0.0.1

# Utilities
aiofiles==23.2.1
pydantic==2.4.2 
